+++ src/dgisim/status/status.py
@@ -334,6 +334,10 @@ __all__ = [
     "RiptideTransferStatus",
     "RiptideStatus",
     "TideWithholderStatus",
+    ## Thoma ##
+    "ASubordinatesSkillsStatus",
+    "BlazingBarrierStatus",
+    "ScorchingOoyoroiStatus",
     ## Tighnari ##
     "KeenSightStatus",
     "VijnanaSuffusionStatus",
@@ -6974,6 +6978,54 @@ class TideWithholderStatus(CharacterHiddenStatus):
                     status=RiptideCounterStatus,
                 ),
             ], None
+        return [], self
+
+
+#### Thoma ####
+
+@dataclass(frozen=True, kw_only=True)
+class ASubordinatesSkillsStatus(TalentEquipmentStatus):
+    @cached_classproperty
+    def CARD(cls) -> type[crd.TalentEquipmentCard]:
+        from ..card.card import ASubordinatesSkills
+        return ASubordinatesSkills
+
+
+@dataclass(frozen=True, kw_only=True)
+class BlazingBarrierStatus(CombatStatus, StackedShieldStatus):
+    usages: int = 1
+    MAX_USAGES: ClassVar[int] = 3
+
+
+@dataclass(frozen=True, kw_only=True)
+class ScorchingOoyoroiStatus(CombatStatus, _UsageStatus):
+    usages: int = 2
+    MAX_USAGES: ClassVar[int] = 2
+    REACTABLE_SIGNALS: ClassVar[frozenset[TriggeringSignal]] = frozenset((
+        TriggeringSignal.POST_SKILL,
+    ))
+
+    @override
+    def _react_to_signal(
+            self, game_state: GameState, source: StaticTarget, signal: TriggeringSignal,
+            detail: None | InformableEvent
+    ) -> tuple[list[eft.Effect], None | Self]:
+        if signal is TriggeringSignal.POST_SKILL:
+            assert isinstance(detail, SkillIEvent)
+            if detail.source.pid is source.pid and detail.skill_true_type.is_normal_attack():
+                return [
+                    eft.ReferredDamageEffect(
+                        source=source,
+                        target=DynamicCharacterTarget.OPPO_ACTIVE,
+                        element=Element.PYRO,
+                        damage=1,
+                        damage_type=DamageType(status=True),
+                    ),
+                    eft.AddCombatStatusEffect(
+                        target_pid=source.pid,
+                        status=BlazingBarrierStatus,
+                    ),
+                ], replace(self, usages=-1)
         return [], self